feat: adding logging, and seperated hetzner code
