+++ command-handler/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
-    "unique-names-generator": "^4.7.1"
+    "unique-names-generator": "^4.7.1",
+    "winston": "^3.14.2"
   }
 }
