+++ command-handler/src/util/logger.js
@@ -1,0 +1,17 @@
+import { createLogger, format, transports } from 'winston';
+
+const { combine, timestamp, json } = format;
+
+export default function logger() {
+    const logger = createLogger({
+        format: combine(
+            timestamp(),
+            json()
+        ),
+        transports: [
+          new transports.Console(),
+        ]
+    });
+
+    return logger;
+}
